# docs/flatmates-openapi.yaml
openapi: 3.1.0
info:
  title: 360 Flatmates API
  description: |
    Complete API for the 360 Flatmates web platform — a swipe-based flatmate-finding app for India's young professionals.

    ## Authentication
    All endpoints require authentication via Bearer token in the `Authorization` header, unless marked as public.
    Tokens are obtained via Supabase Auth (Phone OTP + Password).

    ## User Modes
    - **room_poster** — Has a flat, wants to rent out a spare room
    - **co_hunter** — Looking for people to flat-search with
    - **open_to_both** — Flexible, open to either

    ## Base URL
    `/api/v1`
  version: 1.0.0
  contact:
    name: 360 Ghar
    url: https://360ghar.com
  license:
    name: Confidential
    url: https://360ghar.com
servers:
  - url: https://api.360ghar.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /visits:
    post:
      tags: [Visits]
      summary: Schedule a visit
      description: |
        Schedule a flatmate meet or property tour.
        For flatmate meets, set `visit_context=flatmate_meet` and provide `conversation_id` and `counterparty_user_id`.
      operationId: scheduleVisit
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCreate'
      responses:
        '201':
          description: Visit scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Visits]
      summary: Get my visits
      description: Fetch all visits for the current user.
      operationId: getMyVisits
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of visits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPageVisit'
  /visits/upcoming:
    get:
      tags: [Visits]
      summary: Get upcoming visits
      operationId: getUpcomingVisits
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Upcoming visits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSlice'
  /visits/past:
    get:
      tags: [Visits]
      summary: Get past visits
      operationId: getPastVisits
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Past visits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitSlice'
  /visits/{visit_id}:
    get:
      tags: [Visits]
      summary: Get visit details
      operationId: getVisitDetails
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Visit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Visits]
      summary: Update visit
      description: Update visit status or scheduled date. Used for confirm/reschedule/cancel actions.
      operationId: updateVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitUpdate'
      responses:
        '200':
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
  /visits/{visit_id}/reschedule:
    post:
      tags: [Visits]
      summary: Reschedule a visit
      operationId: rescheduleVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitReschedule'
      responses:
        '200':
          description: Rescheduled visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
  /visits/{visit_id}/cancel:
    post:
      tags: [Visits]
      summary: Cancel a visit
      operationId: cancelVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCancel'
      responses:
        '200':
          description: Cancelled visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
  /visits/{visit_id}/complete:
    post:
      tags: [Visits]
      summary: Mark visit as completed
      description: Mark a visit as completed with optional notes and feedback.
      operationId: completeVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitComplete'
      responses:
        '200':
          description: Completed visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'

  # ===================== LISTINGS =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    VisitCreate:
      type: object
      required: [property_id, scheduled_date]
      properties:
        property_id:
          type: integer
        visit_context:
          $ref: '#/components/schemas/VisitContext'
          default: flatmate_meet
        scheduled_date:
          type: string
          format: date-time
        conversation_id:
          type: integer
          description: Required for flatmate_meet context
        counterparty_user_id:
          type: integer
          description: Required for flatmate_meet context
        match_id:
          type: integer
        special_requirements:
          type: string
        visit_notes:
          type: string
    VisitUpdate:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/VisitStatus'
        scheduled_date:
          type: string
          format: date-time
        special_requirements:
          type: string
        visit_notes:
          type: string
        visitor_feedback:
          type: string
        interest_level:
          type: string
        follow_up_required:
          type: boolean
        follow_up_date:
          type: string
          format: date
    VisitReschedule:
      type: object
      required: [new_date]
      properties:
        new_date:
          type: string
          format: date-time
        reason:
          type: string
    VisitCancel:
      type: object
      properties:
        reason:
          type: string
    VisitComplete:
      type: object
      properties:
        notes:
          type: string
        feedback:
          type: string
    Visit:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        property_id:
          type: integer
        property_title:
          type: string
        agent_id:
          type: integer
        counterparty_user_id:
          type: integer
        conversation_id:
          type: integer
        match_id:
          type: integer
        visit_context:
          $ref: '#/components/schemas/VisitContext'
        scheduled_date:
          type: string
          format: date-time
        actual_date:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/VisitStatus'
        special_requirements:
          type: string
        visit_notes:
          type: string
        visitor_feedback:
          type: string
        interest_level:
          type: string
        follow_up_required:
          type: boolean
        follow_up_date:
          type: string
          format: date
        cancellation_reason:
          type: string
        rescheduled_from:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CursorPageVisit:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Visit'
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        limit:
          type: integer
        total:
          type: integer
          nullable: true
    VisitSlice:
      type: object
      properties:
        visits:
          type: array
          items:
            $ref: '#/components/schemas/Visit'
        total:
          type: integer

    # ===================== LISTING SCHEMAS =====================
...
    VisitContext:
      type: string
      enum: [property_tour, flatmate_meet]
    VisitStatus:
      type: string
      enum: [requested, confirmed, reschedule_suggested, cancelled, completed]
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Optional machine-readable error code